Move PortForward implementation to kubernetes
It's not specific to Openshift, so, it doesn't have to live there.
