+++ kubernetes/kubernetes/src/main/java/org/arquillian/cube/kubernetes/impl/portforward/PortForwardServerConnection.java
@@ -1,4 +1,4 @@
-package org.arquillian.cube.openshift.impl.client;
+package org.arquillian.cube.kubernetes.impl.portforward;
 
 import io.undertow.UndertowMessages;
 import io.undertow.client.ClientCallback;
