+++ openshift/openshift/pom.xml
@@ -31,6 +31,10 @@
 			<groupId>org.jboss.arquillian.core</groupId>
 			<artifactId>arquillian-core-spi</artifactId>
 		</dependency>
+		<dependency>
+			<groupId>org.arquillian.cube</groupId>
+			<artifactId>arquillian-cube-kubernetes</artifactId>
+		</dependency>
 
 		<dependency>
 			<groupId>org.jboss.arquillian.config</groupId>
